feat: completed basic frontend logic
The frontend is able to add user tracks, and receive consumer tracks from the SFU. No styling has been added, so user videos do not display properly. Also there is a bug where user media is not being requested on chrome and the frontend app currently is only working on firefox.

Co-authored-by: Tyler Frye <tyler.frye42@gmail.com>
Co-authored-by: Jay Jayabose <jayabose@gmail.com>