feat: added responsive grid layout and audio toggle
Grid Row/Col calculated and updated dynamically based on the current number of users connected to the call.
Added mic toggling by disabled the state of the audio track, as well as a negotiated data channel to broadcast to other clients that this client is muted. Peer mic status state is visible on the consumers video as a red icon.

Co-authored-by: Stephen Watzman <ixisixi@gmail.com>
Co-authored-by: Jay Jayabose <jayabose@gmail.com>